+++ config/autoload/qr-codes.global.php
@@ -7,6 +7,7 @@ use function Shlinkio\Shlink\Common\env;
 use const Shlinkio\Shlink\DEFAULT_QR_CODE_ERROR_CORRECTION;
 use const Shlinkio\Shlink\DEFAULT_QR_CODE_FORMAT;
 use const Shlinkio\Shlink\DEFAULT_QR_CODE_MARGIN;
+use const Shlinkio\Shlink\DEFAULT_QR_CODE_ROUND_BLOCK_SIZE;
 use const Shlinkio\Shlink\DEFAULT_QR_CODE_SIZE;
 
 return [
@@ -16,6 +17,7 @@ return [
         'margin' => (int) env('DEFAULT_QR_CODE_MARGIN', DEFAULT_QR_CODE_MARGIN),
         'format' => env('DEFAULT_QR_CODE_FORMAT', DEFAULT_QR_CODE_FORMAT),
         'error_correction' => env('DEFAULT_QR_CODE_ERROR_CORRECTION', DEFAULT_QR_CODE_ERROR_CORRECTION),
+        'round_block_size' => (bool) env('DEFAULT_QR_CODE_ROUND_BLOCK_SIZE', DEFAULT_QR_CODE_ROUND_BLOCK_SIZE),
     ],
 
 ];