+++ config/test/test_config.global.php
@@ -8,13 +8,16 @@ use GuzzleHttp\Client;
 use Laminas\ConfigAggregator\ConfigAggregator;
 use Laminas\Diactoros\Response\EmptyResponse;
 use Laminas\ServiceManager\Factory\InvokableFactory;
-use Laminas\Stdlib\Glob;
 use Monolog\Handler\StreamHandler;
 use Monolog\Logger;
 use PHPUnit\Runner\Version;
+use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\ServerRequestInterface;
+use Psr\Http\Server\RequestHandlerInterface;
 use SebastianBergmann\CodeCoverage\CodeCoverage;
 use SebastianBergmann\CodeCoverage\Driver\Selector;
 use SebastianBergmann\CodeCoverage\Filter;
+use SebastianBergmann\CodeCoverage\Report\Html\Facade as Html;
 use SebastianBergmann\CodeCoverage\Report\PHP;
 use SebastianBergmann\CodeCoverage\Report\Xml\Facade as Xml;
 
@@ -27,18 +30,17 @@ use const ShlinkioTest\Shlink\SWOOLE_TESTING_HOST;
 use const ShlinkioTest\Shlink\SWOOLE_TESTING_PORT;
 
 $isApiTest = env('TEST_ENV') === 'api';
-if ($isApiTest) {
+$generateCoverage = env('GENERATE_COVERAGE') === 'yes';
+if ($isApiTest && $generateCoverage) {
     $filter = new Filter();
-    foreach (Glob::glob(__DIR__ . '/../../module/*/src') as $item) {
-        $filter->includeDirectory($item);
-    }
+    $filter->includeDirectory(__DIR__ . '/../../module/Core/src');
+    $filter->includeDirectory(__DIR__ . '/../../module/Rest/src');
     $coverage = new CodeCoverage((new Selector())->forLineCoverage($filter), $filter);
 }
 
 $buildDbConnection = static function (): array {
     $driver = env('DB_DRIVER', 'sqlite');
     $isCi = env('CI', false);
-    $getMysqlHost = static fn (string $driver) => sprintf('shlink_db%s', $driver === 'mysql' ? '' : '_maria');
     $getCiMysqlPort = static fn (string $driver) => $driver === 'mysql' ? '3307' : '3308';
 
     return match ($driver) {
@@ -64,7 +66,7 @@ $buildDbConnection = static function (): array {
         ],
         default => [ // mysql and maria
             'driver' => 'pdo_mysql',
-            'host' => $isCi ? '127.0.0.1' : $getMysqlHost($driver),
+            'host' => $isCi ? '127.0.0.1' : sprintf('shlink_db_%s', $driver),
             'port' => $isCi ? $getCiMysqlPort($driver) : '3306',
             'user' => 'root',
             'password' => 'root',
@@ -114,30 +116,40 @@ return [
 
     'routes' => !$isApiTest ? [] : [
         [
-            'name' => 'start_collecting_coverage',
-            'path' => '/api-tests/start-coverage',
-            'middleware' => middleware(static function () use (&$coverage) {
-                if ($coverage) { // @phpstan-ignore-line
-                    $coverage->start('API tests');
-                }
-                return new EmptyResponse();
-            }),
-            'allowed_methods' => ['GET'],
-        ],
-        [
             'name' => 'dump_coverage',
             'path' => '/api-tests/stop-coverage',
             'middleware' => middleware(static function () use (&$coverage) {
+                // TODO I have tried moving this block to a listener so that it's invoked automatically,
+                //      but then the coverage is generated empty ¯\_(ツ)_/¯
                 if ($coverage) { // @phpstan-ignore-line
                     $basePath = __DIR__ . '/../../build/coverage-api';
-                    $coverage->stop();
+
                     (new PHP())->process($coverage, $basePath . '.cov');
                     (new Xml(Version::getVersionString()))->process($coverage, $basePath . '/coverage-xml');
+                    (new Html())->process($coverage, $basePath . '/coverage-html');
                 }
 
                 return new EmptyResponse();
             }),
             'allowed_methods' => ['GET'],
+        ],
+    ],
+
+    'middleware_pipeline' => !$isApiTest ? [] : [
+        'capture_code_coverage' => [
+            'middleware' => middleware(static function (
+                ServerRequestInterface $req,
+                RequestHandlerInterface $handler,
+            ) use (&$coverage): ResponseInterface {
+                $coverage?->start($req->getHeaderLine('x-coverage-id'));
+
+                try {
+                    return $handler->handle($req);
+                } finally {
+                    $coverage?->stop();
+                }
+            }),
+            'priority' => 9999,
         ],
     ],