+++ module/CLI/src/Command/Tag/ListTagsCommand.php
@@ -45,7 +45,8 @@ class ListTagsCommand extends Command
 
         return map(
             $tags,
-            fn (TagInfo $tagInfo) => [(string) $tagInfo->tag(), $tagInfo->shortUrlsCount(), $tagInfo->visitsCount()],
+            static fn (TagInfo $tagInfo) =>
+                [$tagInfo->tag()->__toString(), $tagInfo->shortUrlsCount(), $tagInfo->visitsCount()],
         );
     }
 }