+++ module/Core/src/Domain/DomainServiceInterface.php
@@ -8,7 +8,6 @@ use Shlinkio\Shlink\Core\Config\NotFoundRedirects;
 use Shlinkio\Shlink\Core\Domain\Model\DomainItem;
 use Shlinkio\Shlink\Core\Entity\Domain;
 use Shlinkio\Shlink\Core\Exception\DomainNotFoundException;
-use Shlinkio\Shlink\Core\Exception\InvalidDomainException;
 use Shlinkio\Shlink\Rest\Entity\ApiKey;
 
 interface DomainServiceInterface
@@ -32,7 +31,6 @@ interface DomainServiceInterface
 
     /**
      * @throws DomainNotFoundException If the API key is restricted to one domain and a different one is provided
-     * @throws InvalidDomainException If default domain is provided
      */
     public function configureNotFoundRedirects(
         string $authority,