+++ module/Core/src/Service/ShortUrl/DeleteShortUrlService.php
@@ -33,7 +33,7 @@ class DeleteShortUrlService implements DeleteShortUrlServiceInterface
         if (! $ignoreThreshold && $this->isThresholdReached($shortUrl)) {
             throw Exception\DeleteShortUrlException::fromVisitsThreshold(
                 $this->deleteShortUrlsOptions->getVisitsThreshold(),
-                $shortUrl->getShortCode(),
+                $identifier,
             );
         }