+++ module/Core/test-db/Domain/Repository/DomainRepositoryTest.php
@@ -50,27 +50,7 @@ class DomainRepositoryTest extends DatabaseTestCase
 
         $this->getEntityManager()->flush();
 
-        self::assertEquals(
-            [$barDomain, $bazDomain, $detachedWithRedirects, $fooDomain],
-            $this->repo->findDomainsWithout(null),
-        );
-        self::assertEquals(
-            [$barDomain, $bazDomain, $detachedWithRedirects],
-            $this->repo->findDomainsWithout('foo.com'),
-        );
-        self::assertEquals(
-            [$bazDomain, $detachedWithRedirects, $fooDomain],
-            $this->repo->findDomainsWithout('bar.com'),
-        );
-        self::assertEquals(
-            [$barDomain, $detachedWithRedirects, $fooDomain],
-            $this->repo->findDomainsWithout('baz.com'),
-        );
-        self::assertEquals(
-            [$barDomain, $bazDomain, $fooDomain],
-            $this->repo->findDomainsWithout('detached-with-redirects.com'),
-        );
-
+        self::assertEquals([$barDomain, $bazDomain, $detachedWithRedirects, $fooDomain], $this->repo->findDomains());
         self::assertEquals($barDomain, $this->repo->findOneByAuthority('bar.com'));
         self::assertEquals($detachedWithRedirects, $this->repo->findOneByAuthority('detached-with-redirects.com'));
         self::assertNull($this->repo->findOneByAuthority('does-not-exist.com'));
@@ -121,14 +101,11 @@ class DomainRepositoryTest extends DatabaseTestCase
 
         $this->getEntityManager()->flush();
 
-        self::assertEquals([$fooDomain], $this->repo->findDomainsWithout(null, $fooDomainApiKey));
-        self::assertEquals([$barDomain], $this->repo->findDomainsWithout(null, $barDomainApiKey));
-        self::assertEquals(
-            [$detachedWithRedirects],
-            $this->repo->findDomainsWithout(null, $detachedWithRedirectsApiKey),
-        );
-        self::assertEquals([$bazDomain, $fooDomain], $this->repo->findDomainsWithout(null, $authorApiKey));
-        self::assertEquals([], $this->repo->findDomainsWithout(null, $authorAndDomainApiKey));
+        self::assertEquals([$fooDomain], $this->repo->findDomains($fooDomainApiKey));
+        self::assertEquals([$barDomain], $this->repo->findDomains($barDomainApiKey));
+        self::assertEquals([$detachedWithRedirects], $this->repo->findDomains($detachedWithRedirectsApiKey));
+        self::assertEquals([$bazDomain, $fooDomain], $this->repo->findDomains($authorApiKey));
+        self::assertEquals([], $this->repo->findDomains($authorAndDomainApiKey));
 
         self::assertEquals($fooDomain, $this->repo->findOneByAuthority('foo.com', $authorApiKey));
         self::assertNull($this->repo->findOneByAuthority('bar.com', $authorApiKey));