+++ module/Rest/src/Action/Domain/ListDomainsAction.php
@@ -7,7 +7,9 @@ namespace Shlinkio\Shlink\Rest\Action\Domain;
 use Laminas\Diactoros\Response\JsonResponse;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
+use Shlinkio\Shlink\Core\Config\NotFoundRedirects;
 use Shlinkio\Shlink\Core\Domain\DomainServiceInterface;
+use Shlinkio\Shlink\Core\Options\NotFoundRedirectOptions;
 use Shlinkio\Shlink\Rest\Action\AbstractRestAction;
 use Shlinkio\Shlink\Rest\Middleware\AuthenticationMiddleware;
 
@@ -16,7 +18,7 @@ class ListDomainsAction extends AbstractRestAction
     protected const ROUTE_PATH = '/domains';
     protected const ROUTE_ALLOWED_METHODS = [self::METHOD_GET];
 
-    public function __construct(private DomainServiceInterface $domainService)
+    public function __construct(private DomainServiceInterface $domainService, private NotFoundRedirectOptions $options)
     {
     }
 
@@ -28,6 +30,7 @@ class ListDomainsAction extends AbstractRestAction
         return new JsonResponse([
             'domains' => [
                 'data' => $domainItems,
+                'defaultRedirects' => NotFoundRedirects::fromConfig($this->options),
             ],
         ]);
     }