+++ module/Rest/test/Action/Domain/ListDomainsActionTest.php
@@ -9,6 +9,7 @@ use Laminas\Diactoros\ServerRequestFactory;
 use PHPUnit\Framework\TestCase;
 use Prophecy\PhpUnit\ProphecyTrait;
 use Prophecy\Prophecy\ObjectProphecy;
+use Shlinkio\Shlink\Core\Config\NotFoundRedirects;
 use Shlinkio\Shlink\Core\Domain\DomainServiceInterface;
 use Shlinkio\Shlink\Core\Domain\Model\DomainItem;
 use Shlinkio\Shlink\Core\Entity\Domain;
@@ -22,11 +23,13 @@ class ListDomainsActionTest extends TestCase
 
     private ListDomainsAction $action;
     private ObjectProphecy $domainService;
+    private NotFoundRedirectOptions $options;
 
     public function setUp(): void
     {
         $this->domainService = $this->prophesize(DomainServiceInterface::class);
-        $this->action = new ListDomainsAction($this->domainService->reveal());
+        $this->options = new NotFoundRedirectOptions();
+        $this->action = new ListDomainsAction($this->domainService->reveal(), $this->options);
     }
 
     /** @test */
@@ -35,7 +38,7 @@ class ListDomainsActionTest extends TestCase
         $apiKey = ApiKey::create();
         $domains = [
             DomainItem::forDefaultDomain('bar.com', new NotFoundRedirectOptions()),
-            DomainItem::forExistingDomain(Domain::withAuthority('baz.com')),
+            DomainItem::forNonDefaultDomain(Domain::withAuthority('baz.com')),
         ];
         $listDomains = $this->domainService->listDomains($apiKey)->willReturn($domains);
 
@@ -46,6 +49,7 @@ class ListDomainsActionTest extends TestCase
         self::assertEquals([
             'domains' => [
                 'data' => $domains,
+                'defaultRedirects' => NotFoundRedirects::fromConfig($this->options),
             ],
         ], $payload);
         $listDomains->shouldHaveBeenCalledOnce();